Refactor `DistributionRout` in `OrderRout`.
 + Renamed `DistributionRout` in `OrderRout` file and method.
 + Modified HTTP server to listen `OrderRout` from `Configuration`.
 + Modified handler to parse in `OrderData` data structure.
 + Project module rename.
