+++ src/orderRout/OrderRout.go
@@ -1,14 +1,14 @@
-package distributionRout
+package orderRout
 
 import (
 	"encoding/json"
-	"github.com/vivk-FAF-PR16-2/RestaurantKitchen/src/utils"
+	"github.com/FAF-PR-RestaurantK/RestaurantKitchen/src/utils"
 	"io"
 	"net/http"
 )
 
-func DistributionHandler(writer http.ResponseWriter, request *http.Request) {
-	var data utils.DistributionData
+func OrderHandler(writer http.ResponseWriter, request *http.Request) {
+	var data utils.OrderData
 	var response string
 
 	jsonData, err := io.ReadAll(request.Body)
